UnrecongizedURLError should be an instanceof Error
