+++ lib/router/transition-intent/url-transition-intent.js
@@ -2,6 +2,7 @@ import TransitionIntent from '../transition-intent';
 import TransitionState from '../transition-state';
 import handlerInfoFactory from '../handler-info/factory';
 import { oCreate, merge, subclass } from '../utils';
+import UnrecognizedURLError from './../unrecognized-url-error';
 
 export default subclass(TransitionIntent, {
   url: null,
@@ -52,13 +53,3 @@ export default subclass(TransitionIntent, {
     return newState;
   }
 });
-
-/**
-  Promise reject reasons passed to promise rejection
-  handlers for failed transitions.
- */
-function UnrecognizedURLError(message) {
-  this.message = (message || "UnrecognizedURLError");
-  this.name = "UnrecognizedURLError";
-}
-
